rename SPORT.{c,h} to sport.{c,h}
