+++ libfreedo/_3do_sys.c
@@ -33,7 +33,7 @@
 #include "Clio.h"
 #include "frame.h"
 #include "Madam.h"
-#include "SPORT.h"
+#include "sport.h"
 #include "XBUS.h"
 #include "DiagPort.h"
 #include "quarz.h"
